+++ blob-store.js
